FIX and WIP: Grafico e front ends
atuliazando que todos 'cursos' tenha uma sala

criacao de MatterController e refatoração de Student para ReponseEntity

+++ src/main/java/io/github/vitinh0z/schoolsystem/controller/StudentController.java
@@ -8,6 +8,7 @@ import io.github.vitinh0z.schoolsystem.repository.ClassRepository;
 import io.github.vitinh0z.schoolsystem.repository.StudentRepository;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.*;
@@ -22,17 +23,21 @@ public class StudentController {
 
 
     @PostMapping
-    public Student createStudent (@Valid @RequestBody Student student){
+    public ResponseEntity<Student> createStudent (@Valid @RequestBody Student student){
 
-        return studentRepository.save(student);
+        return ResponseEntity.ok(studentRepository.save(student));
     }
 
     @PostMapping("/generate")
-    public List<Student> generateStudents (){
+    public ResponseEntity<List<Student>> generateStudents (){
 
         Random random = new Random();
         List<Student> students = new ArrayList<>();
         List<ClassEntity> allClasses = classRepository.findAll();
+
+        if (allClasses.isEmpty()){
+            return ResponseEntity.badRequest().build();
+        }
 
         int i = 0;
         while (i < 300){
@@ -49,43 +54,56 @@ public class StudentController {
 
             i++;
         }
-        return studentRepository.saveAll(students);
+
+        students = studentRepository.saveAll(students);
+
+        return ResponseEntity.ok(students);
     }
 
     @GetMapping("/all")
-    public List<Student> getAllStudents() {
+    public ResponseEntity<List<Student>> getAllStudents() {
 
-        return studentRepository.findAll();
+        return ResponseEntity.ok(studentRepository.findAll());
     }   
 
     @PutMapping("/update/{id}")
-    public String updateStudent(@Valid @RequestParam("id") Integer id,
-                                    @RequestBody Student student){
+    public ResponseEntity<Student> updateStudent( @PathVariable("id") Integer id,
+                                                  @Valid @RequestBody Student studentJson){
 
-        Student findStudent = studentRepository.findById(student.getId()).orElseThrow(()
-                -> new RuntimeException("Estudante não Encotrado")
-        );
+        Optional<Student> find = studentRepository.findById(id);
 
-        findStudent.setName(student.getName());
-        findStudent.setClassId(student.getClassId());
-        studentRepository.save(findStudent);
+        if (find.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
 
-        return "/updateStudent=sucess";
+        Student stundentBanco = find.get();
+
+        stundentBanco.setClassId(studentJson.getClassId());
+        stundentBanco.setName(studentJson.getName());
+
+        Student stundentSalvo = studentRepository.save(stundentBanco);
+
+         return ResponseEntity.ok(stundentSalvo);
     }
 
-    @DeleteMapping("/{id}")
-    public String deleteStudent(@PathVariable("id") Integer id){
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity<Student> deleteStudent(@PathVariable("id") Integer id){
 
-        Student findStudent = studentRepository.findById(id).orElseThrow(()
-                -> new RuntimeException("Estudante não Encotrado")
-        );
+        Optional<Student> findStudent = studentRepository.findById(id);
+
+        if(findStudent.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        Student student = findStudent.get();
 
         studentRepository.deleteById(id);
-        return "/deleteStudent=sucess";
+
+        return ResponseEntity.ok(student);
     }
 
     @GetMapping("dashboard/degree")
-    public List<ClassStatsDTO> getDashboardStatus(){
+    public ResponseEntity<List<ClassStatsDTO>> getDashboardStatus(){
 
         List<Object[]> allStudents = studentRepository.countStudentByDegree();
 
@@ -100,7 +118,7 @@ public class StudentController {
             classList.add(dto);
         }
 
-        return classList;
+        return ResponseEntity.ok(classList);
 
     }
 }
